Move reuse_regrid to module
